@@ -1,11 +1,13 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFrameworks>net5.0;netstandard2.1;</TargetFrameworks>
+    <TargetFramework>net6.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
     <PackageLicenseExpression>MIT</PackageLicenseExpression>
     <PackageProjectUrl></PackageProjectUrl>
     <RepositoryUrl>https://github.com/gdar91/Fills</RepositoryUrl>
-    <Version>0.0.99</Version>
+    <Version>0.0.992</Version>
     <Description>Observable operators, that are missing in the standard library.</Description>
     <Authors>Giorgi Darchiashvili</Authors>
     <Company />
@@ -16,12 +18,8 @@
     <PackageReference Include="System.Reactive" Version="5.0.0" />
   </ItemGroup>
   
-  <ItemGroup Condition="'$(TargetFramework)'=='netstandard2.1'">
-    <PackageReference Include="System.Threading.Channels" Version="5.0.0" />
-  </ItemGroup>
-  
   <ItemGroup>
-    <ProjectReference Include="..\Fills.Hint\Fills.Hint.csproj" />
+    <ProjectReference Include="..\Fills.Core\Fills.Core.csproj" />
   </ItemGroup>
 
 </Project>
