@@ -1,44 +1,42 @@
-using System;
 using System.Reactive.Subjects;
 
-namespace Fills
+namespace Fills;
+
+public sealed class ResettingSubject<T> : ISubject<T>
 {
-    public sealed class ResettingSubject<T> : ISubject<T>
+    private readonly Func<ISubject<T>> subjectFactory;
+
+    private ISubject<T> currentSubject;
+
+
+    public ResettingSubject(Func<ISubject<T>> subjectFactory)
     {
-        private readonly Func<ISubject<T>> subjectFactory;
-
-        private ISubject<T> currentSubject;
-
-
-        public ResettingSubject(Func<ISubject<T>> subjectFactory)
-        {
-            this.subjectFactory = subjectFactory;
-            this.currentSubject = subjectFactory();
-        }
+        this.subjectFactory = subjectFactory;
+        this.currentSubject = subjectFactory();
+    }
 
 
-        public void OnCompleted()
-        {
-            var oldSubject = currentSubject;
-            currentSubject = subjectFactory();
-            oldSubject.OnCompleted();
-        }
+    public void OnCompleted()
+    {
+        var oldSubject = currentSubject;
+        currentSubject = subjectFactory();
+        oldSubject.OnCompleted();
+    }
 
-        public void OnError(Exception error)
-        {
-            var oldSubject = currentSubject;
-            currentSubject = subjectFactory();
-            oldSubject.OnError(error);
-        }
+    public void OnError(Exception error)
+    {
+        var oldSubject = currentSubject;
+        currentSubject = subjectFactory();
+        oldSubject.OnError(error);
+    }
 
-        public void OnNext(T value)
-        {
-            currentSubject.OnNext(value);
-        }
+    public void OnNext(T value)
+    {
+        currentSubject.OnNext(value);
+    }
 
-        public IDisposable Subscribe(IObserver<T> observer)
-        {
-            return currentSubject.Subscribe(observer);
-        }
+    public IDisposable Subscribe(IObserver<T> observer)
+    {
+        return currentSubject.Subscribe(observer);
     }
 }
